* Added address of node to operation error description as optional
