+++ internal/xerrors/operation_test.go
@@ -34,6 +34,10 @@ func Test_operationError_Error(t *testing.T) {
 		text string
 	}{
 		{
+			err:  Operation(WithStatusCode(Ydb.StatusIds_BAD_REQUEST), WithNodeAddress("localhost")),
+			text: "operation/BAD_REQUEST (code = 400010, address = localhost)",
+		},
+		{
 			err:  Operation(WithStatusCode(Ydb.StatusIds_BAD_REQUEST)),
 			text: "operation/BAD_REQUEST (code = 400010)",
 		},
